Reformat, fix RichParser suggestions

+++ worldedit-core/src/main/java/com/sk89q/worldedit/extension/factory/parser/mask/AdjacentMaskParser.java
@@ -14,12 +14,12 @@ import java.util.stream.Stream;
 public class AdjacentMaskParser extends RichParser<Mask> {
 
     public AdjacentMaskParser(WorldEdit worldEdit) {
-        super(worldEdit,"~");
+        super(worldEdit, "~");
     }
 
     @Override
     protected Stream<String> getSuggestions(String argumentInput, int index) {
-        if(index == 0){
+        if (index == 0) {
             return worldEdit.getMaskFactory().getSuggestions(argumentInput).stream();
         } else if (index == 1 || index == 2) {
             return this.suggestPositiveDoubles(argumentInput);
@@ -29,7 +29,7 @@ public class AdjacentMaskParser extends RichParser<Mask> {
 
     @Override
     protected Mask parseFromInput(@NotNull String[] arguments, ParserContext context) throws InputParseException {
-        if(arguments.length == 0) return null;
+        if (arguments.length == 0) return null;
         Mask subMask = worldEdit.getMaskFactory().parseFromInput(arguments[0], context);
         int min = arguments.length > 1 ? Integer.parseInt(arguments[1]) : -1;
         int max = arguments.length > 2 ? Integer.parseInt(arguments[2]) : -1;
@@ -40,6 +40,6 @@ public class AdjacentMaskParser extends RichParser<Mask> {
         if (max >= 8 && min == 1) {
             return new AdjacentAnyMask(subMask);
         }
-        return new AdjacentMask(subMask, (int) min, (int) max);
+        return new AdjacentMask(subMask, min, max);
     }
 }
